// Patientor/Patientor/src/types.ts
export interface Diagnosis {
  code: string;
  name: string;
  latin?: string;
}

export enum Gender {
  Male = "Male",
  Female = "Female",
  Other = "Other"
}

export interface Patient {
  id: string;
  name: string;
  occupation: string;
  gender: Gender;
  ssn?: string;
  dateOfBirth?: string;
  entries: Entry[];
}

interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Array<Diagnosis['code']>;
}

export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}

export interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}

interface SickLeave {
  startDate: string;
  endDate: string;
}

export interface OccupationalHealthCareEntry extends BaseEntry {
  type: "OccupationalHealthCare";
  employerName: string;
  sickLeave?: SickLeave;
}

interface Discharge {
  date: string;
  criteria: string;
}

export interface HospitalEntry extends BaseEntry {
  type: "Hospital";
  discharge: Discharge;
}

export type Entry = 
| HospitalEntry
| OccupationalHealthCareEntry
| HealthCheckEntry;

export interface PatientWithEntry extends Patient{
  entries: Entry[]
}

export type EntryFormValues = 
| Omit<HospitalEntry, 'id'>
| Omit<OccupationalHealthCareEntry, 'id'>
| Omit<HealthCheckEntry, 'id'>;

export enum Type {
  Hospital = "Hospital",
  OccupationalHealthCare = "OccupationalHealthCare",
  HealthCheck = "HealthCheck",
}